fix(renovate): schedule on the first day of the month
The monthly schedule preset translates to `before 3am on the first day of the month`. That being before 3am was causing issues with the updates as it takes some time for Renovate to run and the checks to be passed before creating a PR. Therefore, the schedule causes only the branches to be created and no PR's. Setting the schedule to the complete first day should solve this issue.