ci(release): increase build workflow to 15 minutes
GitHub API requests can easy take a couple of seconds and the release can make lots of API requests
when a release has a lot of commits. If every requests takes five seconds, which is on the high
side for a request, this could perform 180 requests. This should be enough for most use cases.